Parse file length as int64 in gridfs and update tests to use DownloadToStream.
GODRIVER-654

+++ mongo/gridfs/bucket.go
@@ -336,7 +336,7 @@ func (b *Bucket) openDownloadStream(filter interface{}, opts ...*options.FindOpt
 		return nil, err
 	}
 
-	fileLen := fileLenElem.Int32()
+	fileLen := fileLenElem.Int64()
 	if fileLen == 0 {
 		return newDownloadStream(nil, b.chunkSize, 0), nil
 	}
